Use dynamic domain

+++ entry.sh
@@ -1,14 +1,22 @@
 #!/bin/sh
 set -e
 
-darkhttpd /root/letsencrypt/www/ --port 80 --daemon
+DOMAIN=${DOMAIN:-}
 
-# Ensure we run latest version of acme.sh
-curl https://get.acme.sh | sh
-# ls /root/.acme.sh
-/root/.acme.sh/acme.sh  --issue  -d jeroen.intercity.io -w /root/letsencrypt/www --staging \
- --debug --standalone --accountkey /root/letsencrypt/account.key \
- --certhome /root/letsencrypt/certs/
+if [ "${DOMAIN}" ]
+then
+  echo "Doing stuff ${DOMAIN}"
+  darkhttpd /root/letsencrypt/www/ --port 80 --daemon
 
-rm -rf /root/letsencrypt/www
+  # Ensure we run latest version of acme.sh
+  curl https://get.acme.sh | sh
+  # ls /root/.acme.sh
+  /root/.acme.sh/acme.sh  --issue  -d $DOMAIN -w /root/letsencrypt/www --staging \
+   --debug --standalone --accountkey /root/letsencrypt/account.key \
+   --certhome /root/letsencrypt/certs/
+
+  rm -rf /root/letsencrypt/www
+else
+  echo "Domain is required"
+fi
